backend/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==2.3.2
flask-cors
jsonlines==3.1.0
transformers==4.25.1
torch==1.13.0
gunicorn==20.1.0
openai~=1.38.0
python-dotenv~=1.0.0
nltk~=3.8.1
pip~=21.3.1
wheel~=0.37.1
MarkupSafe~=2.1.3
Werkzeug~=2.3.6
click~=8.1.6
blinker~=1.6.2
Jinja2~=3.1.2
itsdangerous~=2.1.2
setuptools~=60.2.0
zipp~=3.16.2
tenacity~=8.2.2
azure-cognitiveservices-speech
google-generativeai
anthropic
mistralai
pyyaml